rename eventExplore widget file
