+++ lib/screens/explorePage.dart
@@ -1,6 +1,5 @@
 import 'package:events_app/widgets/customtext.dart';
-import 'package:events_app/widgets/event_explore.dart';
-import 'package:events_app/widgets/event_feed.dart';
+import 'package:events_app/widgets/eventExplore_widget.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
