+++ lib/widgets/eventExplore_widget.dart
